refactor: separate general metrics tooling from objective 4
